Clean up: rename rss.ts to blog.ts, remove Leaflet RSS-specific styles
- Rename lib/rss.ts → lib/blog.ts and update all imports
- Remove unused .postContent, .listMarker, and flex-based li overrides
  that were specific to Leaflet's RSS HTML structure
- Simplify list styles for our own generated HTML
